Feature/fix/modulo reservas

+++ src/main/java/com/sanfrancisco/api/modules/servicios/dto/response/ServicioResponse.java
@@ -8,7 +8,7 @@ public record ServicioResponse(
         Integer tipoServicioId,
         String tipoServicioNombre,
         Integer estanciaId,
-        BigDecimal cantidad,
+        Integer cantidad,
         BigDecimal precioAplicado,
         BigDecimal subtotal,
         String observaciones,
